+++ CHANGELOG.md
@@ -19,7 +19,7 @@
 ### Docs ###
 
 * Add multiple inline docs
-* Add [Example](https://memeozer.com/) Website to showcase most of the features
+* Add [Example](https://uximprovements.memeozer.com/) Website to showcase most of the features
 
 ## 0.11.5